Reorganized tests a bit.
